+++ source-code/core/src/utilities/types.ts
@@ -1,0 +1,1 @@
+export type MaybePromise<T> = T | Promise<T>;
